feat : /api/record/puuid 조회시 데이터 응답 형태 리팩토링 + 소환사 아이콘, 레벨 조회

+++ src/main/java/com/lolprojectbackend/record/dto/SummonerInfoDto.java
@@ -9,7 +9,7 @@ public class SummonerInfoDto {
     private String id;  // encryptedSummonerId
     private String accountId;
     private String puuid;
-    private int profileIconId;
+    private Long profileIconId;
     private long revisionDate;
-    private int summonerLevel;
+    private Long summonerLevel;
 }
